Backend/requirements.txt
fastapi
starlette
pydantic
python-dotenv
uvicorn[standard]
groq
pdf2image
Pillow
python-multipart
